Убрал подчёркивание у ссылок. Переименовал Layout в MessagesScreen.
