+++ lesson5/src/routes/Error_404/index.js
@@ -3,7 +3,12 @@ import { Box } from "@mui/material";
 export const Error_404 = () => {
   return (
     <Box
-      sx={{ display: "flex", justifyContent: "center", alignItems: "center" }}
+      sx={{
+        width: "100%",
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+      }}
     >
       <h1>Error 404. Page not found</h1>
     </Box>
