+++ lesson5/src/routes/Main/index.js
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 import "./App.css";
 import { Paper } from "@mui/material";
-import { Layout } from "../../Components/Layout";
+import { Layout } from "../../Components/MessagesScreen";
 import { Chats } from "../../Components/Chats";
 import { Header } from "../../Components/Header";
 import { Search } from "../../Components/Search";
